updated changes
2. Either add a "category" column to the do-not-call XLS list or use
separate worksheets in the workbook
a. The apache poi can use separate worksheets very easily
b. The do not call xls in my branch is updated, will need to be put in
folder when binary .exe is made
3. Deal with the fact that one exon can have many variants with
potentially different do-not-call statuses
a. Only flashes Warning if all variants on do not call
b. Else no warning but user will see on do not call when look at variant
list
4. Add a column to the variant list within the exon that represents the
do-not-call status
a. Done
5. Consider backing the do-not-call list with an entity class and using
a static populate method (optional)
a. I tried my best to attempt to do this, Do Not Call is its own entity
class using populate but maybe because I’m not 100% clear on shared
variables in java, I still am passing in representation of Do Not Call
(this time as an arraylist instead of string[][]) when populating
variant objects, so not sure if my implementation is really more modular
but I think it may be easier to follow (especially since I got rid of
string[][] representation)

1. Automatically making a copy of the tsv file as an excel file
a. Per Dr. Mann she is requiring us to do this for every case
2. In excel sheet highlighting variable that are on the do not call or
are synonymous
a. Highlight variables means ultimately they we not called when signing
out the case
b. Ideally should probably create by comparing original with variants
selected for import but I feel that may take a little more time (and I
want to get some of this pushed to the live version)